refactor: separate bluetooth logic and add MAX30101 init
