+++ src/MAX30101.h
@@ -1,0 +1,6 @@
+#ifndef MAX30101_H_
+#define MAX30101_H_
+
+int max30101_init(void);
+
+#endif /* MAX30101_H_ */
